+++ ImageFeed/Presentations/Authentification/WebViewViewController.swift
@@ -47,34 +47,29 @@ final class WebViewViewController: UIViewController {
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        
-//        webView.addObserver(
-//            self,
-//            forKeyPath: #keyPath(WKWebView.estimatedProgress),
-//            options: .new,
-//            context: nil)
-        
         updateProgress()
-        
     }
     
-    override func viewDidDisappear(_ animated: Bool) {
-        super.viewDidDisappear(animated)
-//        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress), context: nil)
-    }
+    //    func startActivity() {
+    //    }
+    
+    //    override func viewDidDisappear(_ animated: Bool) {
+    //        super.viewDidDisappear(animated)
+    ////        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress), context: nil)
+    //    }
     
     // MARK: - Overrides Methods
-//    override func observeValue(
-//        forKeyPath keyPath: String?,
-//        of object: Any?,
-//        change: [NSKeyValueChangeKey : Any]?,
-//        context: UnsafeMutableRawPointer?) {
-//            if keyPath == #keyPath(WKWebView.estimatedProgress) {
-//                updateProgress()
-//            } else {
-//                super.observeValue(forKeyPath: keyPath, of: object, change: change, context: context)
-//            }
-//        }
+    //    override func observeValue(
+    //        forKeyPath keyPath: String?,
+    //        of object: Any?,
+    //        change: [NSKeyValueChangeKey : Any]?,
+    //        context: UnsafeMutableRawPointer?) {
+    //            if keyPath == #keyPath(WKWebView.estimatedProgress) {
+    //                updateProgress()
+    //            } else {
+    //                super.observeValue(forKeyPath: keyPath, of: object, change: change, context: context)
+    //            }
+    //        }
     
     // MARK: - IB Actions
     @IBAction func didTappBackButton(_ sender: Any) {
@@ -119,7 +114,18 @@ extension WebViewViewController: WKNavigationDelegate {
         {
             return codeItem.value
         } else {
+//            showAlertInWebView()
             return nil
         }
+    }
+    
+    private func showAlertInWebView() {
+        let alertController = UIAlertController(
+            title: "Что-то пошло не так(",
+            message: "Не удалось войти в систему",
+            preferredStyle: .alert)
+        let cancelAction = UIAlertAction(title: "ОК", style: .cancel)
+        alertController.addAction(cancelAction)
+        present(alertController, animated: true)
     }
 }
